// src/shared/core/enums/permission-code.enum.ts
export enum PermissionCode {

  PROJECT_READ = "PROJECT_READ",
  PROJECT_CREATE = "PROJECT_CREATE",
  PROJECT_UPDATE = "PROJECT_UPDATE",
  PROJECT_DELETE = "PROJECT_DELETE",

  TASK_READ = "TASK_READ",
  TASK_CREATE = "TASK_CREATE",
  TASK_UPDATE = "TASK_UPDATE",
  TASK_DELETE = "TASK_DELETE",

  WORKFLOW_READ = "WORKFLOW_READ",
  WORKFLOW_UPDATE = "WORKFLOW_UPDATE",
  WORKFLOW_REVIEW = "WORKFLOW_REVIEW",

  MASTER_DATA_READ = "MASTER_DATA_READ",
  MASTER_DATA_UPDATE = "MASTER_DATA_UPDATE",

  USER_READ = "USER_READ",
  USER_CREATE = "USER_CREATE",
  USER_UPDATE = "USER_UPDATE",
  USER_DELETE = "USER_DELETE",
  USER_PERMISSION_OVERRIDE_MANAGE = "USER_PERMISSION_OVERRIDE_MANAGE",

  COMMENT_READ = "COMMENT_READ",
  COMMENT_CREATE = "COMMENT_CREATE",
  COMMENT_UPDATE = "COMMENT_UPDATE",
  COMMENT_DELETE = "COMMENT_DELETE",
  COMMENT_DELETE_ANY = "COMMENT_DELETE_ANY",

  ACTIVITY_LOG_READ = "ACTIVITY_LOG_READ",
  ACTIVITY_LOG_CREATE = "ACTIVITY_LOG_CREATE",
  ACTIVITY_LOG_READ_ANY = "ACTIVITY_LOG_READ_ANY",
  ACTIVITY_LOG_DELETE = "ACTIVITY_LOG_DELETE",
  ACTIVITY_TYPE_MANAGE = "ACTIVITY_TYPE_MANAGE",

  ROLE_MANAGE = "ROLE_MANAGE",

  NESTING_READ = "NESTING_READ"

}